feat: `POST /profile-img/form` 삭제 후 기능 MQ로 이전
main에 있던 REDIS_AUTH 추가 로직 삭제
profile-img.service 삭제 후일 새로 추가 예정
