+++ src/profile-img/profile-img.controller.ts
@@ -1,5 +1,4 @@
 import {
-  Body,
   Controller,
   Post,
   UploadedFile,
@@ -7,19 +6,12 @@ import {
 } from '@nestjs/common';
 import { FileInterceptor } from '@nestjs/platform-express/multer';
 import { ApiTags } from '@nestjs/swagger';
-import { ProfileImgService } from './profile-img.service';
 import { TempImageStorage } from '@src/common/storage/multer-s3.storage';
-import {
-  FileLocationResultDto,
-  ProfileFormDto,
-  ProfileFormResultDto,
-} from '@src/dto/upload.dto';
+import { FileLocationResultDto } from '@src/dto/upload.dto';
 
 @ApiTags('profile-img')
 @Controller('profile-img')
 export class ProfileImgController {
-  constructor(private readonly uploadService: ProfileImgService) {}
-
   @Post('temp')
   @UseInterceptors(
     FileInterceptor('images', {
@@ -30,10 +22,5 @@ export class ProfileImgController {
     @UploadedFile() file: Express.MulterS3.File,
   ): FileLocationResultDto {
     return { location: file.location, key: file.key };
-  }
-
-  @Post('form')
-  uploadForm(@Body() dto: ProfileFormDto): Promise<ProfileFormResultDto> {
-    return this.uploadService.toProfile(dto);
   }
 }
