+++ src/profile-img/profile-img.module.ts
@@ -1,9 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ProfileImgController } from './profile-img.controller';
-import { ProfileImgService } from './profile-img.service';
 
 @Module({
   controllers: [ProfileImgController],
-  providers: [ProfileImgService],
 })
 export class ProfileImgModule {}
